1. 去掉Unsafe 2. 调整模版
