+++ core/src/main/resources/com/github/ompc/greys/core/res/javascript/greys-module.js
