+++ core/src/main/resources/com/github/ompc/greys/core/res/version
@@ -1,1 +1,1 @@
-1.7.4.0
+1.7.5.0
