+++ core/src/test/java/com/github/ompc/greys/core/js/JavaScriptTestCase.java
@@ -4,6 +4,7 @@ import com.github.ompc.greys.core.util.GaStringUtils;
 import org.apache.commons.io.IOUtils;
 import org.junit.After;
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
 
 import javax.script.*;
@@ -14,6 +15,7 @@ import java.nio.charset.Charset;
  * JavaScript测试用例
  * Created by vlinux on 16/2/16.
  */
+@Ignore
 public class JavaScriptTestCase {
 
     protected ScriptEngine scriptEngine;
@@ -33,11 +35,15 @@ public class JavaScriptTestCase {
      * 加载JavaScriptSupport
      */
     private void loadJavaScriptSupport() throws IOException, ScriptException {
-        final String javascriptSupportContent = IOUtils.toString(
-                GaStringUtils.class.getResourceAsStream("/com/github/ompc/greys/core/res/javascript/javascript-support2.js"),
+        compilable.compile(IOUtils.toString(
+                GaStringUtils.class.getResourceAsStream("/com/github/ompc/greys/core/res/javascript/javascript-support.js"),
                 Charset.forName("UTF-8")
-        );
-        compilable.compile(javascriptSupportContent).eval();
+        )).eval();
+
+        compilable.compile(IOUtils.toString(
+                GaStringUtils.class.getResourceAsStream("/com/github/ompc/greys/core/res/javascript/greys-module.js"),
+                Charset.forName("UTF-8")
+        )).eval();
     }
 
     @After
@@ -48,10 +54,10 @@ public class JavaScriptTestCase {
     }
 
     @Test
-    public void test_compile_javascript_support_success() throws IOException, ScriptException {
+    public void test_compile_javascript_support_success() throws IOException, ScriptException, NoSuchMethodException {
 
-
-
+        invocable.invokeFunction("__greys_load", "/tmp/logger.js", "UTF-8");
+        invocable.invokeFunction("__greys_module_returning", null,null,null);
 
     }
