+++ scripts/module_template.js
@@ -1,0 +1,28 @@
+/**
+ * 模版
+ */
+__greys_define(['greys'], function () {
+    greys.watching({
+
+        create: function (output) {
+
+        },
+
+        destroy: function (output) {
+
+        },
+
+        before: function (output, advice) {
+
+        },
+
+        returning: function (output, advice) {
+
+        },
+
+        throwing: function (output, advice) {
+
+        },
+
+    });
+})
